requirements.txt
streamlit 
tensorflow 
streamlit-drawable-canvas 
Pillow 
numpy
opencv-python